move enum to type package
